refactor(offerings, validation): add field to model and remove some functions from validation
Enrolment code may be needed, so it is relevant to add it (May remove
the whole offering model later as I think we might not actually need it,
we may just need the OfferingRequest model)

Remove some sanitization fucntions since sanitizing is not needed,
golang escapes strings by default when adding them to the database. We
may need to scape html tags or find a way to block users fron adding
them as classroom/exercie names on the frontend though.