refactor(routes): fixes code splitting of the routes
By adding the nav and footer inside de App component, it made so that
the main js file after build got too bloated. By adding the Nav inside
the pages component we make so that the bundler can split the code in a
smarter way getting smaller js files.

Also, the login page wont have a nav and footter, so it is good that
they are not global